autonomous-it/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Kubernetes client
kubernetes==28.1.1

# HTTP client for Telegram
httpx==0.25.2

# Configuration
pydantic-settings==2.1.0

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Utilities
python-json-logger==2.0.7
structlog==23.2.0
tenacity==8.2.3
